Changed observation_block to observation_request. This will now hold a function to get the target of an observation request. Updated DB to hold target coordinate values (right ascension, declination). Data generator now makes better coordinates.
