+++ db/createDB.sql
@@ -12,7 +12,8 @@ DROP TABLE IF EXISTS observation;
 
 CREATE TABLE observation (
   id bigint(20) NOT NULL ,
-  sky_address varchar(191) NOT NULL,
+  right_ascension float(6,3) NOT NULL,
+  declination float(6,3) NOT NULL,
   agency_id bigint(20) NOT NULL,
   priority bigint(10) NOT NULL,
   remaining_observing_chances integer(10) NOT NULL,
